+++ spec/lib/bright_serializer/serializer_spec.rb
@@ -80,6 +80,33 @@ RSpec.describe BrightSerializer::Serializer do
       end
 
       it { expect(instance.to_hash).to eq(result) }
+
+      context 'when the object is an Hash with string keys' do
+        let(:serializer_class) do
+          Class.new do
+            include BrightSerializer::Serializer
+            attributes :first_name, :last_name
+          end
+        end
+
+        let(:user_hash) do
+          {
+            'first_name' => user.first_name,
+            'last_name' => user.last_name
+          }
+        end
+
+        let(:result) do
+          {
+            first_name: user.first_name,
+            last_name: user.last_name
+          }
+        end
+
+        it 'serialize all 3 attributes' do
+          expect(serializer_class.new(user_hash).to_hash).to eq(result)
+        end
+      end
     end
   end